+++ src/posthog-core.js
@@ -245,7 +245,7 @@ PostHogLib.prototype._init = function (token, config, name) {
 
     this['_jsc'] = function () {}
 
-    this._captureMetrics = new CaptureMetrics(this.get_config('_capture_metrics'), _.bind(this.capture, this))
+    this._captureMetrics = new CaptureMetrics(this.get_config('_capture_metrics'))
 
     this._requestQueue = new RequestQueue(this._captureMetrics, _.bind(this._handle_queued_event, this))
 
@@ -379,7 +379,6 @@ PostHogLib.prototype._handle_unload = function () {
     if (this.get_config('_capture_metrics')) {
         this._requestQueue.updateUnloadMetrics()
         this.capture('$capture_metrics', this._captureMetrics.metrics)
-        this._captureMetrics.captureInProgressRequests()
     }
     this._requestQueue.unload()
     this._retryQueue.unload()
@@ -795,10 +794,7 @@ PostHogLib.prototype.reloadFeatureFlags = function () {
  *                              It'll return a list of feature flags enabled for the user.
  */
 PostHogLib.prototype.onFeatureFlags = function (callback) {
-    this.featureFlags.addFeatureFlagsHandler(callback)
-    const flags = this.featureFlags.getFlags()
-    const flagVariants = this.featureFlags.getFlagVariants()
-    callback(flags, flagVariants)
+    this.featureFlags.onFeatureFlags(callback)
 }
 
 /**